sml calculations integrated but tests failing
