+++ src/main/scala/uk/gov/ons/registers/model/CommonFrameDataFields.scala
@@ -1,0 +1,18 @@
+package uk.gov.ons.registers.model
+
+object CommonFrameDataFields {
+  val prn = "prn"
+  val cellNumber = "cell_no"
+  val payeEmployees = "paye_empees"
+  val sic07 = "sic07"
+
+  val employees = "paye_empees"
+  val jobs = "paye_jobs"
+  val ern = "ern"
+
+  val contained = "cntd_turnover"
+  val apportioned = "app_turnover"
+  val standard = "std_turnover"
+  val group_turnover = "grp_turnover"
+  val ent = "ent_turnover"
+}
