+++ src/test/resources/data/calculations/README
@@ -147,3 +147,21 @@ Calculations
              grp_turnover = 1000
              ent_turnover = 260 + 444 = 704
              contd_turnover = null
+
++----------+-----+-------+------+------+--------+-----+-------------------+
+| PAYE REF | dec | march | june | sept |  total | avg | last quorter emps |
++----------+-----+-------+------+------+--------+-----+-------------------+
+|   1152L  |   5 |   6   |   -  |   8  |   19   |  6  |         8         |
++----------+-----+-------+------+------+--------+-----+-------------------+
+|   1153L  |   9 |   1   |   2  |   3  |   15   |  3  |         3         |
++----------+-----+-------+------+------+--------+-----+-------------------+
+|   1154L  |   4 |   -   |   6  |   7  |   17   |  5  |         7         |
++----------+-----+-------+------+------+--------+-----+-------------------+
+|   1155L  |   8 |   9   |   1  |   2  |   20   |  5  |         2         |
++----------+-----+-------+------+------+--------+-----+-------------------+
+
+                                    paye_empees:  19
+                                                    paye_jobs: 20
+
+
+                               actual:            17           20
