+++ src/test/scala/closures/AddNewPeriodWithCalculationsSpec.scala
@@ -45,15 +45,15 @@ class AddNewPeriodWithCalculationsSpec extends HBaseConnectionManager with Paths
 
   }
 
-  "dummy tests" should{
+/*  "dummy tests" should{
 
     "create report files to make Jenkins happy" in{
       true shouldBe true
     }
 
-  }
+  }*/
 
-/* val appConfs = AppParams(
+ val appConfs = AppParams(
    (Array[String](
      "LINKS", "ons", "l", linkHfilePath,
      "LEU", "ons", "d", leuHfilePath,
@@ -69,7 +69,7 @@ class AddNewPeriodWithCalculationsSpec extends HBaseConnectionManager with Paths
      "add-calculated-period"
    )))
 
-  override d beforeAll() = {
+  override def beforeAll() = {
    implicit val spark: SparkSession = SparkSession.builder().master("local[4]").appName("enterprise assembler").getOrCreate()
    conf.set("hbase.zookeeper.quorum", "localhost")
    conf.set("hbase.zookeeper.property.clientPort", "2181")
@@ -81,6 +81,17 @@ class AddNewPeriodWithCalculationsSpec extends HBaseConnectionManager with Paths
     }
    spark.stop
 }
+
+
+/*  override def afterAll() = {
+    File(parquetPath).deleteRecursively()
+    File(linkHfilePath).deleteRecursively()
+    File(leuHfilePath).deleteRecursively()
+    File(entHfilePath).deleteRecursively()
+    File(louHfilePath).deleteRecursively()
+    File(ruHfilePath).deleteRecursively()
+    File(existingRecordsDir).deleteRecursively()
+  }*/
 
 /*  "blaCree test-data csvh" should {"blah" in{
        implicit val spark: SparkSession = SparkSession.builder().master("local[4]").appName("enterprise assembler").getOrCreate()
@@ -176,7 +187,7 @@ def createRecords(appconf:AppParams)(implicit spark: SparkSession,connection:Con
   saveToHFile(existingRusForNewPeriodScenario,appconf.HBASE_REPORTINGUNITS_COLUMN_FAMILY, appconf, existingRusRecordHFiles)
   saveToHFile(existingLeusForNewPeriodScenario,appconf.HBASE_ENTERPRISE_COLUMN_FAMILY, appconf, existingLeusRecordHFiles)
   saveToHFile(existingEntsForNewPeriodScenario,appconf.HBASE_ENTERPRISE_COLUMN_FAMILY, appconf, existingEntRecordHFiles)
-}*/
+}
 
 
 }
